fix: Unify tool definitions between tools/list and MCP server routes
This change ensures that tools/list command returns the same schema as what's
used by the MCP server routes:

1. Created a shared module for tool definitions (dynamic-definitions.ts)
2. Updated tools/list and setupRoutes to both use this shared module
3. Ensures parameters consistency when environmnet variables or CLI
   options are provided

Now tool parameters will be correctly visible regardless of how
the server was started.
